requirements.txt
discord.py==2.1.0
names==0.3.0
Pillow==9.4.0
